pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markdown-code-runner"
version = "0.1.0"
description = "Execute and optionally rewrite code blocks in Markdown files based on external commands"
authors = [{ name = "Pol Dellaiera", email = "pol.dellaiera@protonmail.com" }]
readme = "README.md"
license = { text = "EUPL-1.2" }
requires-python = ">=3.12"
dependencies = [
    "markdown-it-py>=3.0.0",
    "pydantic-settings>=2.8.1"
]
[project.scripts]
mdcr = "markdown_code_runner.main:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
